ml/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
xgboost
tensorflow
statsmodels
prophet
transformers
torch
shap
requests
python-multipart
joblib
yfinance
newsapi-python
langchain
langchain-google-genai
google-generativeai
pandas-ta